// src/widgets/detailInfoWidget/models/Detail.ts
export interface DetailInfo {
	id: number
	name: string
	url: string
	created: string
	status?: string
	species?: string
	type?: string
	gender?: string
	origin?: {}
	location?: {}
	image?: string
	episode?: string[]
	dimension?: string
	residents?: string[]
	air_date?: string
	characters?: string[]
}